+++ recipes/install-custom-metrics.rb
@@ -56,6 +56,13 @@ cookbook_file "jvm_metrics.sh" do
   mode "755"
 end
 
+cookbook_file "num_online_workers.sh" do
+  path "/usr/local/bin/num_online_workers.sh"
+  owner "root"
+  group "root"
+  mode "755"
+end
+
 cron_d 'disk_metrics' do
   user 'custom_metrics'
   minute '*/2'
@@ -122,6 +129,15 @@ if mh_node?
     user "root"
     minute "*/2"
     command %Q(/usr/local/bin/jvm_metrics.sh "#{aws_instance_id}" felix 2>&1 | logger -t info)
+    path '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin'
+  end
+end
+
+if admin_node?
+  cron_d 'online_workers_metrics' do
+    user 'custom_metrics'
+    minute '*/2'
+    command %Q(/usr/local/bin/num_online_workers.sh "#{stack_id}" 2>&1 | logger -t info)
     path '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin'
   end
 end